Add Book() and addBookToLibrary() functions
In this commit, I have added an object constructor to create a new
object containing the title, author, number of pages, and read status of
a book. From there, I added an array and a function to push each
constructed book object to the array, as well as console.log() lines to
test this as I continue developing the script.