Fix errors, edit title
I corrected the function connecting the form values to the variables;
now, what used to be the addBookToLibrary() function is now a variable
calling the object constructor function, so as to ensure that the form
values are actually used as the form variables based on scope. I also
corrected the page title within the HTML head.